extras: Makefile in separate clauses; comment on get_bit func
Skeptical about function: get_bit should it be:
    bits & (1 << i)  vs current (bits >> i) & 0x01
but added comment for resolution by debate.
Introduced the usage of a Makefile. By default, running make
creates geohash.o. make test compiles into a binary and actually
runs the test.
