+++ .gitignore
@@ -1,0 +1,4 @@
+*.o
+*~
+*.sw[op]
+*_test
